Creating the controller for the User class.

+++ core/src/main/java/springsql/core/us/userUS/DeleteUserByIdUS.java
@@ -9,6 +9,9 @@ public record DeleteUserByIdUS(UserRepositoryPort userRepositoryPort) implements
 
     @Override
     public void apply(String id) {
-        userRepositoryPort.deleteById(id);
+        if (!userRepositoryPort.existsById(id))
+            throw new RuntimeException("User not found.");
+        else
+            userRepositoryPort.deleteById(id);
     }
 }
